+++ internal/cabinet/config/decor.go
@@ -13,6 +13,8 @@ var ValidDecorThemeIDs = []string{
 	"pink",
 	"orange",
 	"yellow",
+	"violet",
+	"slate",
 	"neon",
 	"new_year",
 	"summer",
@@ -20,6 +22,11 @@ var ValidDecorThemeIDs = []string{
 	"valentine",
 	"spring",
 	"black_friday",
+	"aurora",
+	"ocean",
+	"cyber",
+	"sunset",
+	"lavender",
 }
 
 var validDecorThemes map[string]struct{}